server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
pydantic>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
